Removes ReductionRunUtils
